Add S3 creation flow draft 6 - rename s3 configuration class
