+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_data_source_configuration/direct_query_amazon_s3_datasource/direct_query_configure_amazon_s3_data_sources.tsx
